Introduce bust: a Rust-like language for type checking exploration (#26)
* Add bust language scaffold with grammar, tokens, and stub lexer

New language project with Rust-like syntax targeting HM type inference.
Includes initial grammar spec, TokenType enum, token aliases, lexer
factory with stub implementation, and REPL entry point.


* Implement bust lexer to hopefully match the grammar I'm outlining

* Add lexer tests, comment grammar rules, and CTest target for bust

Lexer tests cover single/double char tokens, keywords, identifiers,
integer literals (including large numbers stored as strings), comments,
whitespace, and compound sequences like let bindings and function defs.


* Add better printing of Tokens across core

* Improve token_type_to_string to show enum name with lexeme

Format: COMMA ',', ARROW "->", LET "let". Single quotes for
single-char tokens, double quotes for multi-char. IDENTIFIER and
INT_LITERAL stay plain since they carry their lexeme in the token value.


* Add stub parser with peek/advance/expect token helpers


* Add stub ast.hpp and include from parser


* First pass at ast node design

* Add parser method stubs and comprehensive parser test suite

Stubs for all grammar rules in the precedence chain. Tests cover
literals, identifiers, function defs, let bindings, arithmetic with
precedence and associativity, comparison, logical ops, unary, function
calls, if/else, return, blocks, lambdas, and a fibonacci integration
test.


* Almost done parsing

* Add AST dumper and DUMP_AST to all parser tests


* Add ENABLE_ASAN option for AddressSanitizer support


* Parse lambdas, parse expressions not requiring semicolons after them

* Add grammar rule for block-like statements without semicolons

Block-ending expressions (if, bare blocks, while, for) can be used
as statements without a trailing semicolon.


* Change SourceLocation to hold a const char* rather than a std::string, so it cannot throw (bad_alloc in case of string)

* Disable bugprone-unchecked-optional-access in test directories

DocTest REQUIRE guards optional access but clang-tidy does not
see through the macro, causing false positives.


* Add readability-braces-around-statements to clang-tidy


* Fix braces-around-statements warnings in ast_dump.hpp


* Move programs/ into blip/ since they are blip-specific


---------

+++ core/inc/token.hpp
@@ -27,8 +27,7 @@ public:
   virtual ~Token() = default;
 
   virtual std::ostream &dump(std::ostream &out) const {
-    out << get_location()
-        << " TokenType: " << static_cast<unsigned int>(get_token_type());
+    out << get_location() << " TokenType: " << get_token_type();
     return out;
   }
